Cleaned up file io code
