+++ src/lib.rs
@@ -7,97 +7,6 @@ mod window;
 pub mod audio;
 pub mod config;
 
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-
-#[derive(Serialize, Deserialize)]
-pub struct Config {
-    /// The name of the ball to use for this experient
-    pub active_ball: Option<String>,
-
-    /// Where the output data gets stored to once the experiment is done
-    pub output_data_path: Option<String>,
-
-    pub balls: Vec<Ball>,
-    pub dials: Vec<Dial>,
-    pub alarms: Vec<Alarm>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct Ball {
-    pub name: String,
-    /// The action that this ball takes.
-    ///
-    /// Can be one of `random_direction`, `random_speed`, or `random_direction_change`
-    pub action: BallAction,
-
-    /// How quickly a ball configured for `random_direction_change` changes speed.
-    pub random_direction_change_time: Option<f64>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct Dial {
-    /// The name of the alarm this dial uses
-    pub alarm: String,
-
-    pub start: f32,
-    pub end: f32,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct Alarm {
-    /// The user defined name of this alarm. Used to match up which alarm is being used in
-    /// [`Dial::alarm`]
-    pub name: String,
-
-    /// The path to the audio file for this alarm
-    pub audio_path: String,
-
-    /// The key that clears this alarm.
-    /// Case insensitive single letter
-    pub clear_key: char,
-}
-
-pub enum BallAction {
-    /// The ball moves at a random speed
-    RandomSpeed,
-
-    /// The ball moves in a random direction from the crosshair
-    RandomDirection,
-
-    /// The ball changes directions randomly in a certain interval
-    RandomDirectionChange,
-}
-
-impl Serialize for BallAction {
-    fn serialize<S>(&self, se: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let s = match self {
-            BallAction::RandomSpeed => "random_direction",
-            BallAction::RandomDirection => "random_speed",
-            BallAction::RandomDirectionChange => "random_direction_change",
-        };
-        s.serialize(se)
-    }
-}
-
-impl<'de> Deserialize<'de> for BallAction {
-    fn deserialize<D>(de: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let s = String::deserialize(de)?;
-        match s.as_str() {
-            "random_direction" => Ok(BallAction::RandomDirection),
-            "random_speed" => Ok(BallAction::RandomSpeed),
-            "random_direction_change" => Ok(BallAction::RandomDirectionChange),
-            _ => Err(serde::de::Error::custom(format!(
-                "Unknown ball action `{s}`"
-            ))),
-        }
-    }
-}
 const DEFAULT_INPUT_PATH: &str = "./config.toml";
 
 pub fn run() {
@@ -105,46 +14,14 @@ pub fn run() {
         Ok(toml) => match toml::from_str(&toml) {
             Ok(t) => t,
             Err(e) => {
-                panic!();
+                println!("failed to parse config file");
+                println!("{}", e);
+                std::process::exit(1);
             }
         },
-        /*for i in 0..num_dials {
-            let dial = Dial::new(
-                i + 1,
-                50.0,
-                DialRange::new(i as f32 * 200.0, i as f32 * 200.0 + range_size),
-                (i + 1).to_string().chars().next().unwrap(),
-            );
-            dials.push(dial);
-        }*/
         Err(_) => {
-            let range_size = 4000.0;
-            let config = Config {
-                balls: vec![Ball {
-                    name: "Ball 1".to_owned(),
-                    action: BallAction::RandomSpeed,
-                    random_direction_change_time: Some(4.0),
-                }],
-                dials: (1u32..=5)
-                    .map(|i| Dial {
-                        start: i as f32 * 200.0,
-                        end: i as f32 * 200.0 + range_size,
-                        alarm: i.to_string(),
-                    })
-                    .collect(),
-
-                alarms: (1u32..=5)
-                    .map(|i| Alarm {
-                        name: i.to_string(),
-                        audio_path: "alarm.wav".to_owned(),
-                        clear_key: char::from_digit(i, 10).unwrap(),
-                    })
-                    .collect(),
-
-                active_ball: Some("Ball 1".to_owned()),
-                output_data_path: None,
-            };
             // Write out default config if none existed before
+            let config = config::Config::default();
             let toml = toml::to_string(&config).unwrap();
             std::fs::write(DEFAULT_INPUT_PATH, &toml).unwrap();
 
@@ -156,7 +33,7 @@ pub fn run() {
     window::draw_gui(&config);
 }
 
-fn validate_config(config: &mut Config) {
+fn validate_config(config: &mut config::Config) {
     if let Some(active) = &config.active_ball {
         let ball_names: Vec<_> = config.balls.iter().map(|b| &b.name).collect();
         if !ball_names.contains(&active) {
@@ -165,6 +42,9 @@ fn validate_config(config: &mut Config) {
             std::process::exit(1);
         }
     }
+
+    // We need to print `alarm_names` in the event of an error
+    #[allow(clippy::needless_collect)]
     let alarm_names: Vec<_> = config.alarms.iter().map(|b| &b.name).collect();
     for dial in &config.dials {
         let alarm_name = &dial.alarm;
